--- REST_Microservices/AboService/src/main/java/com/micro/shop/consumer/imported/Product.java
@@ -1,23 +1,0 @@
-package com.micro.shop.consumer.imported;
-
-public class Product {
-
-    private Integer nr;
-    private String name;
-    private Float price;
-
-    public Product() {
-    }
-
-    public Integer getNr() {
-        return nr;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Float getPrice() {
-        return price;
-    }
-}
