+++ REST_Microservices/CartService/pom.xml
@@ -51,6 +51,14 @@
             <version>3.0.0</version>
         </dependency>
         <!-- ∧ Swagger API Documentation -->
+        <!-- ⋁ Lombok to skip boilerplate code -->
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.24</version>
+            <scope>provided</scope>
+        </dependency>
+        <!-- ∧ Lombok to skip boilerplate code -->
     </dependencies>
 
     <build>
