+++ REST_Microservices/CartService/src/main/java/com/micro/api/CartService.java
@@ -1,4 +1,4 @@
-package com.micro.shop;
+package com.micro.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
