+++ REST_Microservices/CartService/src/main/java/com/micro/api/external/IAboService.java
@@ -1,0 +1,5 @@
+package com.micro.api.external;
+
+public interface IAboService {
+    void addAbo(Integer productNr);
+}
