+++ REST_Microservices/CartService/src/main/java/com/micro/api/external/IProductService.java
@@ -1,0 +1,5 @@
+package com.micro.api.external;
+
+public interface IProductService {
+    Product getProduct(Integer productNr);
+}
