--- REST_Microservices/CartService/src/main/java/com/micro/shop/consumer/imported/ProductSelector.java
@@ -1,5 +1,0 @@
-package com.micro.shop.consumer.imported;
-
-public interface ProductSelector {
-    Product getProduct(Integer productNr);
-}
