+++ REST_Microservices/CartService/src/test/java/com/micro/api/CartServiceTests.java
@@ -1,4 +1,4 @@
-package com.micro.shop;
+package com.micro.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
